refactor(app): extract comparison page
- Move Agent comparison logic into a dedicated page
- Keep the application entry focused on page composition
